move operator/* to controller/
