+++ go/controller/config_test.go
@@ -3,7 +3,7 @@ package operator_test
 import (
 	"testing"
 
-	"github.com/PaddlePaddle/cloud/go/operator"
+	"github.com/PaddlePaddle/cloud/go/controller"
 	"github.com/stretchr/testify/assert"
 )
